Add styles to likes and hobbies section
